Start Summaries & a little change In SuccessContract's Channel Eevent

+++ Application/BackgroundWorker/Common/Events/SuccessedContractUpdated.cs
@@ -4,6 +4,6 @@ namespace Application.BackgroundWorker.Common.Events
 {
     public class SuccessedContractUpdated
     {
-        public int SuccessedContractId { get; set; }
+        public long SuccessedContractId { get; set; }
     }
 }
